Creation of the smartasphalt package
Created a directory to act as a location to store files that are related to the smart asphalt project as a package and moved key files their. Also update main to reflect the use of the package in the import statements. Other files left out of the package were not dependent on the key pieces of software developed for the project.