Creation of setup.py for package initialization
Created a simple skeleton for a setup.py file for packages that we create within our project. Currently just has a basic outline of the package and executes find_packages() to look for pakcages in the project. On a seperate note, the files generated from executing the command "python3 setup.py develop" will be ignored and not committed to the repo.