Iteration of Cruise Controller with Hall Effect Encoder
Utilizing the Halleffect encoder from the Arduino we will measure velocity using millimeters per second as of now. Given that measured velocity that will be passed into the new cruise control based PID controller to hopefully attain a desired velocity.
Currently their is some untested code that needs to be completed and verified for the PI cruise controller.