Updated the Velocity Controller Integral Term
For the integral term it is calculated using integral_term * accumulated_error * change_in_time. Previously the change in time was not considered, however with advice from a colleague in cubesat change_in_time is equivalenty to 1/sampling_rate which is what is used here, sampling rate is 150 samples per minute or 1 sample every 400 millisecs which is acquired from looking at the arduino implementation. This is approximately 2.5Hz frequency.