Network Follower Implementation with Lidar
Implemented the network follower by sharing the velocity which is fed into the existing control system design, does not utilize the cruise controller but the other lidar/encoder pid controller. Will need to implement the sharing of the servo position from leader to follower however steering control system is hard to determine how it takes in a desired steering value from the leader.