Updates: Platoon Leader and Controls
Changed the Platoon leader logic to be easier to follow and set/get data in a effective order. Controls had minor changes one being that when a new setpoint is assigned the accumulated and prev_error are set to 0. Also I added some nice discrete PID controller implmentation examples. Created a file called path planner that is meant to execute a specified set of commands within a time interval this is different than vehicle command script and directly interfaces with Platoon Leader. Debug tools is a file added to help create commands for debugging purposes.